feat(cangku): 实现供应商管理功能
- 新增供应商分页查询接口，支持按编码、名称、联系人等条件搜索
- 实现供应商创建、更新、删除及状态变更功能- 添加供应商编码和名称唯一性校验逻辑
- 完善供应商相关请求和响应实体类定义- 补充服务层分页查询及数据转换逻辑- 增加控制器层统一异常处理机制

+++ my-project-backend/src/main/java/com/example/entity/cangku/dto/Inventory.java
@@ -6,6 +6,8 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import com.example.entity.dto.BaseModel;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 //库存表
 @Data
 @TableName("ck_inventory")
@@ -19,7 +21,7 @@ public class Inventory extends BaseModel {
     //产品ID
     private Long productId;
     //当前库存数量
-    private Integer quantity;
+    private BigDecimal quantity;
     //锁定数量
-    private Integer lockedQuantity;
+    private BigDecimal lockedQuantity;
 }
